requirements.txt
alembic            1.13.1
apkInspector       1.2.4
appdirs            1.4.4
asn1crypto         1.5.1
asttokens          2.4.1
banal              1.0.6
cffi               1.16.0
click              8.1.7
colorama           0.4.6
contourpy          1.2.1
cycler             0.12.1
dataset            1.6.2
decorator          5.1.1
dpkt               1.9.8
exceptiongroup     1.2.1
executing          2.0.1
fonttools          4.53.0
frida              16.3.3
greenlet           3.0.3
ipython            8.25.0
jedi               0.19.1
joblib             1.4.2
kiwisolver         1.4.5
loguru             0.7.2
lxml               5.2.2
Mako               1.3.5
MarkupSafe         2.1.5
matplotlib         3.9.0
matplotlib-inline  0.1.7
mutf8              1.0.6
networkx           3.3
nfstream           6.5.3
numpy              1.26.4
oscrypto           1.3.0
packaging          24.0
pandas             2.2.2
parso              0.8.4
pefile             2023.2.7
pillow             10.3.0
pip                24.0
prompt_toolkit     3.0.46
psutil             5.9.8
pure-eval          0.2.2
pycparser          2.22
pydot              2.0.0
Pygments           2.18.0
pyparsing          3.1.2
PySide6            6.7.1
PySide6_Addons     6.7.1
PySide6_Essentials 6.7.1
python-dateutil    2.9.0.post0
pytz               2024.1
PyYAML             6.0.1
scapy              2.5.0
scikit-learn       1.5.0
scipy              1.13.1
setuptools         70.0.0
six                1.16.0
SQLAlchemy         1.4.52
stack-data         0.6.3
termcolor          2.4.0
threadpoolctl      3.5.0
traitlets          5.14.3
typing_extensions  4.12.2
tzdata             2024.1
watchdog           4.0.1
wcwidth            0.2.13
win32-setctime     1.1.0
yara-python        4.5.1